Fix test helpers after media aggregator changes
